upgrade

+++ system/controllers/register.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  *  PHP Mikrotik Billing (https://github.com/hotspotbilling/phpnuxbill/)
  *  by https://t.me/ibnux
@@ -10,21 +11,21 @@ if (isset($routes['1'])) {
     $do = 'register-display';
 }
 
-$otpPath = 'system/cache/sms/';
+$otpPath = $CACHE_PATH . File::pathFixer('/sms/');
 
 switch ($do) {
     case 'post':
         $otp_code = _post('otp_code');
-        $username = alphanumeric(_post('username'),"+_.");
+        $username = alphanumeric(_post('username'), "+_.@-");
         $email = _post('email');
         $fullname = _post('fullname');
         $password = _post('password');
         $cpassword = _post('cpassword');
         $address = _post('address');
-        if(!empty($config['sms_url'])){
+        if (!empty($config['sms_url']) && $_c['allow_phone_otp'] == 'yes') {
             $phonenumber = Lang::phoneFormat($username);
             $username = $phonenumber;
-        }else if(strlen($username)<21){
+        } else if (strlen($username) < 21) {
             $phonenumber = $username;
         }
         $msg = '';
@@ -41,46 +42,44 @@ switch ($do) {
             $msg .= 'Email is not Valid<br>';
         }
         if ($password != $cpassword) {
-            $msg .= $_L['PasswordsNotMatch'] . '<br>';
+            $msg .= Lang::T('Passwords does not match') . '<br>';
         }
 
-        if(!empty($config['sms_url'])){
-            $otpPath .= sha1($username.$db_password).".txt";
+        if (!empty($config['sms_url']) && $_c['allow_phone_otp'] == 'yes') {
+            $otpPath .= sha1($username . $db_pass) . ".txt";
             run_hook('validate_otp'); #HOOK
             //expired 10 minutes
-            if(file_exists($otpPath) && time()-filemtime($otpPath)>1200){
+            if (file_exists($otpPath) && time() - filemtime($otpPath) > 1200) {
                 unlink($otpPath);
                 r2(U . 'register', 's', 'Verification code expired');
-            }else if(file_exists($otpPath)){
+            } else if (file_exists($otpPath)) {
                 $code = file_get_contents($otpPath);
-                if($code!=$otp_code){
+                if ($code != $otp_code) {
                     $ui->assign('username', $username);
                     $ui->assign('fullname', $fullname);
                     $ui->assign('address', $address);
                     $ui->assign('email', $email);
                     $ui->assign('phonenumber', $phonenumber);
-                    $ui->assign('notify', '<div class="alert alert-success">
-                    <button type="button" class="close" data-dismiss="alert">
-                    <span aria-hidden="true">×</span>
-                    </button>
-                    <div>Verification code is Wrong</div></div>');
-                    $ui->display('register-otp.tpl');
+                    $ui->assign('notify', 'Wrong Verification code');
+                    $ui->assign('notify_t', 'd');
+                    $ui->assign('_title', Lang::T('Register'));
+                    $ui->display('user-ui/register-otp.tpl');
                     exit();
-                }else{
+                } else {
                     unlink($otpPath);
                 }
-            }else{
+            } else {
                 r2(U . 'register', 's', 'No Verification code');
             }
         }
         $d = ORM::for_table('tbl_customers')->where('username', $username)->find_one();
         if ($d) {
-            $msg .= $_L['account_already_exist'] . '<br>';
+            $msg .= Lang::T('Account already axist') . '<br>';
         }
         if ($msg == '') {
             run_hook('register_user'); #HOOK
             $d = ORM::for_table('tbl_customers')->create();
-            $d->username = alphanumeric($username,"+_.");
+            $d->username = alphanumeric($username, "+_.@-");
             $d->password = $password;
             $d->fullname = $fullname;
             $d->address = $address;
@@ -88,20 +87,18 @@ switch ($do) {
             $d->phonenumber = $phonenumber;
             if ($d->save()) {
                 $user = $d->id();
-                r2(U . 'login', 's', $_L['Register_Success']);
+                r2(U . 'login', 's', Lang::T('Register Success! You can login now'));
             } else {
                 $ui->assign('username', $username);
                 $ui->assign('fullname', $fullname);
                 $ui->assign('address', $address);
                 $ui->assign('email', $email);
                 $ui->assign('phonenumber', $phonenumber);
-                $ui->assign('notify', '<div class="alert alert-danger">
-                <button type="button" class="close" data-dismiss="alert">
-                <span aria-hidden="true">×</span>
-                </button>
-                <div>Failed to register</div></div>');
+                $ui->assign('notify', 'Failed to register');
+                $ui->assign('notify_t', 'd');
+                $ui->assign('_title', Lang::T('Register'));
                 run_hook('view_otp_register'); #HOOK
-                $ui->display('register-rotp.tpl');
+                $ui->display('user-ui/register-rotp.tpl');
             }
         } else {
             $ui->assign('username', $username);
@@ -109,61 +106,57 @@ switch ($do) {
             $ui->assign('address', $address);
             $ui->assign('email', $email);
             $ui->assign('phonenumber', $phonenumber);
-            $ui->assign('notify', '<div class="alert alert-danger">
-            <button type="button" class="close" data-dismiss="alert">
-            <span aria-hidden="true">×</span>
-            </button>
-            <div>' . $msg . '</div></div>');
-            $ui->display('register.tpl');
+            $ui->assign('notify', $msg);
+            $ui->assign('notify_t', 'd');
+            $ui->assign('_title', Lang::T('Register'));
+            $ui->display('user-ui/register.tpl');
         }
         break;
 
     default:
-        if(!empty($config['sms_url'])){
+        if (!empty($config['sms_url']) && $_c['allow_phone_otp'] == 'yes') {
             $username = _post('username');
-            if(!empty($username)){
+            if (!empty($username)) {
                 $d = ORM::for_table('tbl_customers')->where('username', $username)->find_one();
                 if ($d) {
-                    r2(U . 'register', 's', $_L['account_already_exist']);
+                    r2(U . 'register', 's', Lang::T('Account already axist'));
                 }
-                if(!file_exists($otpPath)){
+                if (!file_exists($otpPath)) {
                     mkdir($otpPath);
-                    touch($otpPath.'index.html');
+                    touch($otpPath . 'index.html');
                 }
-                $otpPath .= sha1($username.$db_password).".txt";
+                $otpPath .= sha1($username . $db_pass) . ".txt";
                 //expired 10 minutes
-                if(file_exists($otpPath) && time()-filemtime($otpPath)<1200){
+                if (file_exists($otpPath) && time() - filemtime($otpPath) < 600) {
                     $ui->assign('username', $username);
-                    $ui->assign('notify', '<div class="alert alert-success">
-                    <button type="button" class="close" data-dismiss="alert">
-                    <span aria-hidden="true">×</span>
-                    </button>
-                    <div>Please wait '.(1200-(time()-filemtime($otpPath))).' seconds before sending another SMS</div></div>');
-                    $ui->display('register-otp.tpl');
-                }else{
-                    $otp = rand(100000,999999);
+                    $ui->assign('notify', 'Please wait ' . (600 - (time() - filemtime($otpPath))) . ' seconds before sending another SMS');
+                    $ui->assign('notify_t', 'd');
+                    $ui->assign('_title', Lang::T('Register'));
+                    $ui->display('user-ui/register-otp.tpl');
+                } else {
+                    $otp = rand(100000, 999999);
                     file_put_contents($otpPath, $otp);
-                    Message::sendSMS($username,$config['CompanyName']."\nYour Verification code are: $otp");
+                    Message::sendSMS($username, $config['CompanyName'] . "\n\n".Lang::T("Registration code")."\n$otp");
                     $ui->assign('username', $username);
-                    $ui->assign('notify', '<div class="alert alert-success">
-                    <button type="button" class="close" data-dismiss="alert">
-                    <span aria-hidden="true">×</span>
-                    </button>
-                    <div>Verification code has been sent to your phone</div></div>');
-                    $ui->display('register-otp.tpl');
+                    $ui->assign('notify', 'Registration code has been sent to your phone');
+                    $ui->assign('notify_t', 's');
+                    $ui->assign('_title', Lang::T('Register'));
+                    $ui->display('user-ui/register-otp.tpl');
                 }
-            }else{
+            } else {
+                $ui->assign('_title', Lang::T('Register'));
                 run_hook('view_otp_register'); #HOOK
-                $ui->display('register-rotp.tpl');
+                $ui->display('user-ui/register-rotp.tpl');
             }
-        }else{
+        } else {
             $ui->assign('username', "");
             $ui->assign('fullname', "");
             $ui->assign('address', "");
             $ui->assign('email', "");
             $ui->assign('otp', false);
+            $ui->assign('_title', Lang::T('Register'));
             run_hook('view_register'); #HOOK
-            $ui->display('register.tpl');
+            $ui->display('user-ui/register.tpl');
         }
         break;
 }
